unify SchedulerLogParser and MapReduceContainerLogParser run method logic to common text parser.
