+++ task-parser/src/main/java/com/oppo/cloud/parser/service/job/parser/MapReduceContainerLogParser.java
@@ -16,21 +16,12 @@
 
 package com.oppo.cloud.parser.service.job.parser;
 
-import com.oppo.cloud.common.constant.ProgressState;
-import com.oppo.cloud.common.domain.job.LogPath;
 import com.oppo.cloud.common.util.textparser.ParserAction;
-import com.oppo.cloud.parser.domain.job.CommonResult;
 import com.oppo.cloud.parser.domain.job.ParserParam;
-import com.oppo.cloud.parser.domain.reader.ReaderObject;
-import com.oppo.cloud.parser.service.reader.IReader;
-import com.oppo.cloud.parser.service.reader.LogReaderFactory;
-import com.oppo.cloud.parser.service.writer.OpenSearchWriter;
 import com.oppo.cloud.parser.service.writer.ParserResultSink;
 import lombok.extern.slf4j.Slf4j;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 @Slf4j
 public class MapReduceContainerLogParser extends CommonTextParser {
@@ -39,45 +30,6 @@ public class MapReduceContainerLogParser extends CommonTextParser {
                                        List<ParserAction> actions,
                                        ParserResultSink parserResultSink) {
         super(param, actions, parserResultSink);
-    }
-
-    @Override
-    public CommonResult run() {
-        CommonResult commonResult = new CommonResult();
-        commonResult.setLogType(this.param.getLogType());
-        List<String> categories = new ArrayList<>();
-        updateParserProgress(ProgressState.PROCESSING, 0, 0);
-        String logType = this.param.getLogType().getName();
-
-        for (LogPath logPath : this.param.getLogPaths()) {
-            List<ReaderObject> readerObjects;
-            try {
-                IReader reader = LogReaderFactory.create(logPath);
-                readerObjects = reader.getReaderObjects();
-            } catch (Exception e) {
-                log.error("Exception:", e);
-                continue;
-            }
-            updateParserProgress(ProgressState.PROCESSING, 0, readerObjects.size());
-            for (ReaderObject readerObject : readerObjects) {
-                Map<String, ParserAction> results;
-                try {
-                    results = super.parse(readerObject);
-                } catch (Exception e) {
-                    log.error("Exception:", e);
-                    continue;
-                } finally {
-                    readerObject.close();
-                }
-                List<String> list = getSink().saveParserActions(
-                        logType, readerObject.getLogPath(), this.param, results);
-                categories.addAll(list);
-            }
-        }
-
-        commonResult.setResult(categories);
-        updateParserProgress(ProgressState.SUCCEED, 0, 0);
-        return commonResult;
     }
 
 }
